upgrade
Former-commit-id: dce8cadf09e40928f719251dae5f146032f08f9f

+++ dawdler/dawdler-core/src/main/java/com/anywide/dawdler/core/net/buffer/DirectBufferCreator.java
@@ -41,7 +41,7 @@ public class DirectBufferCreator implements BufferCreator {
 		try {
 			unsafe = Unsafe.getUnsafe();
 			Class<?> bufferClass = Class.forName("java.nio.DirectByteBuffer");
-			bufferConstructor = bufferClass.getDeclaredConstructor(long.class, int.class);
+			bufferConstructor = bufferClass.getDeclaredConstructor(long.class, long.class);
 			bufferConstructor.setAccessible(true);
 		} catch (Exception e) {
 			logger.error("", e);
